Remove legacy weirdness in Merge::Downsample
In practice, this will have only marginal effect. The length_limit
was increased from 6.7 ms to 10 ms. This is compared with the
input_length, which is equal to the decoded frame size. Thus,
this change will only affect encoded frame sizes in this range
(including 10 ms).

BUG=2696
R=tina.legrand@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/9969004