GN: Fix webrtc/video/BUILD.gn for Chromium build.
A mistake was made in https://review.webrtc.org/18709004/
so it doesn't build in Chromium. Adding a config to get
the root folder included in the include path solves it.

BUG=3441
TESTED=Local compilation of Chromium's all target with
src/third_party/webrtc linked to the WebRTC checkout with
this CL applied.
TBR=brettw@chromium.org

Review URL: https://webrtc-codereview.appspot.com/19169004